ml_service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
pdfplumber==0.9.0
